Rename "Configuration" to "Options", small improvements.
To avoid confusion with Zend\Config, renamed all classes and tests to Options.
 * renamed Configuration* to Options*
 * improved fromArray()
 * $ignoreUnknown flag is true by default
